backend/requirements.txt
langchain
langchain-openai
openai
fastapi
uvicorn
python-dotenv
pydantic
